backend/requirements.txt
python = "^3.12"
fastapi = "0.104.1"
llama-index = "0.9.19"
pypdf = "3.17.0"
python-dotenv = "1.0.0"
uvicorn = {version = "0.23.2", extras = ["standard"]}
pymongo = "^4.10.1"